src/pyproject.toml
[project]
name = "spotify-mcp"
version = "0.2.0"
description = "MCP spotify project"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
 "mcp>=1.8.0",
 "python-dotenv>=1.0.1",
 "spotipy==2.24.0",
]

[[project.authors]]
name = "Varun Srivastava"
email = "varun.neal@berkeley.edu"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "typer>=0.9.0",
]

[project.scripts]
spotify-mcp = "server:main"
